requirements.txt
matplotlib~=3.6.1
tqdm~=4.64.1
speedtest-cli~=2.1.3
